# [clang-tidy] Improve `bugprone-exception-escape`: add stacktrace of escaped exception