--- a/clang-tools-extra/test/clang-tidy/checkers/bugprone/exception-escape-rethrow.cpp
+++ b/clang-tools-extra/test/clang-tidy/checkers/bugprone/exception-escape-rethrow.cpp
@@ -1,4 +1,4 @@
-// RUN: %check_clang_tidy -std=c++11,c++14,c++17,c++20 %s bugprone-exception-escape %t -- \
+// RUN: %check_clang_tidy -std=c++11-or-later %s bugprone-exception-escape %t -- \
 // RUN:     -- -fexceptions
 
 void rethrower() {
@@ -22,8 +22,7 @@ int throwsAndCallsRethrower() noexcept {
     }
     return 1;
 }
-// CHECK-MESSAGES: :[[@LINE-9]]:5: note: throw stack of unhandled exception, starting from function 'throwsAndCallsRethrower'
-// CHECK-MESSAGES: :[[@LINE-7]]:9: note: frame #0: function 'throwsAndCallsRethrower' throws unhandled exception here
+// CHECK-MESSAGES: :[[@LINE-6]]:9: note: frame #0: unhandled exception may be thrown in function 'throwsAndCallsRethrower' here
 
 int throwsAndCallsCallsRethrower() noexcept {
 // CHECK-MESSAGES: :[[@LINE-1]]:5: warning: an exception may be thrown in function 'throwsAndCallsCallsRethrower' which should not throw exceptions
@@ -34,8 +33,7 @@ int throwsAndCallsCallsRethrower() noexcept {
     }
     return 1;
 }
-// CHECK-MESSAGES: :[[@LINE-9]]:5: note: throw stack of unhandled exception, starting from function 'throwsAndCallsCallsRethrower'
-// CHECK-MESSAGES: :[[@LINE-7]]:9: note: frame #0: function 'throwsAndCallsCallsRethrower' throws unhandled exception here
+// CHECK-MESSAGES: :[[@LINE-6]]:9: note: frame #0: unhandled exception may be thrown in function 'throwsAndCallsCallsRethrower' here
 
 void rethrowerNoexcept() noexcept {
     throw;