… clang-tools-extra/test/clang-tidy/checkers/bugprone/exception-escape-coro.cpp
… clang-tools-extra/test/clang-tidy/checkers/bugprone/exception-escape-throw.cpp
… clang-tools-extra/test/clang-tidy/checkers/bugprone/exception-escape.cpp