This PR add stacktrace of escaped exception to `bugprone-exception-escape` check.
Changes:
1. Modified `ExceptionAnalyzer` and `ExceptionInfo` classes to hold stacktrace of escaped exception in `llvm::MapVector`. `llvm::MapVector` is needed to hold relative positions of functions in stack as well as have fast lookup.
2. Added new diagnostics based of `misc-no-recursion` check.

**Example of new diagnostics:**

> warning: an exception may be thrown in function 'calls_non_and_throwing' which should not throw exceptions [bugprone-exception-escape]
note: frame #0: unhandled exception may be thrown in function 'explicit_int_thrower' here
note: frame #1: function 'indirect_explicit' calls function 'explicit_int_thrower' here

More example can be found it tests.

**Performance downgrade:**
Run tests on [poco](https://github.com/pocoproject/poco) library, which is known for using exceptions. 679 files total.

**Before:**
`time` command

> real	64m27.072s
user	64m21.617s
sys	0m4.221s

summed-up `--enable-check-profile` command
> Total wall time: 39.796020 seconds
Total user time: 39.146827 seconds
Total system time: 0.633553 seconds

**After:**
`time` command

> real	65m20.295s
user	65m14.240s
sys	0m4.413s

summed-up `--enable-check-profile` command
> Total wall time: 42.621696 seconds
Total user time: 41.897438 seconds
Total system time: 0.705210 seconds

Fixes https://github.com/llvm/llvm-project/issues/87422