@vbvictor 2025-04-10
Some more real-world examples how it was before
![image](https://github.com/user-attachments/assets/7fe19331-2baa-4615-90e1-6c763ad8179f)
and after
![image](https://github.com/user-attachments/assets/970b387c-83c2-4569-b752-c5d66d46e237)