@vbvictor 2025-04-16
Added handling of unknown exceptions, now we diag on place where `throw(int) `is written.
Most of the comments in https://reviews.llvm.org/D153298 are covered in my PR.