@vbvictor 2025-06-18
> Could you add some separate tests for this diagnostic?

I've added tests with:
- with template functions
- deeper nesting
- multiple exceptions throw (as for now, it reports for only one, and I'd like to leave it for first iteration)

If there are any more tests that I should, please tell me.