@vbvictor 2025-07-12
Ran this check on `Poco`, `VTK` and `opencv` successfully, stack traces are printed correctly.